--- pessoa.py
@@ -1,14 +1,0 @@
-"""
-Classe Pessoa define o tipo de dado que será manipulado na memória RAM e Cache
-"""
-
-class Pessoa:
-    def __init__(self, nome, telefone, endereco, aniversario):
-        self.nome = nome
-        self.telefone = telefone
-        self. endereco = endereco
-        self.aniversario = aniversario
-    
-    def informacoes(self):
-        print(f'\nNome: {self.nome}\nTelefone: {self.telefone}\nEndereco: {self.endereco}\nAniversario: {self.aniversario}\n')
-
